Move xtask::gencomp module to its own file
